moved query index generation into a seprate component
