+++ bin/lib/components/index.js
@@ -8,4 +8,6 @@ var LinearCombination_1 = require("./LinearCombination");
 exports.LinearCombination = LinearCombination_1.LinearCombination;
 var LowDegreeProver_1 = require("./LowDegreeProver");
 exports.LowDegreeProver = LowDegreeProver_1.LowDegreeProver;
+var QueryIndexGenerator_1 = require("./QueryIndexGenerator");
+exports.QueryIndexGenerator = QueryIndexGenerator_1.QueryIndexGenerator;
 //# sourceMappingURL=index.js.map
