+++ bin/lib/utils/index.js
@@ -2,7 +2,6 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 // IMPORTS
 // ================================================================================================
-const crypto = require("crypto");
 const inliners = require("./inliners");
 // RE-EXPORTS
 // ================================================================================================
@@ -29,35 +28,6 @@ function isPowerOf2(value) {
     }
 }
 exports.isPowerOf2 = isPowerOf2;
-function getPseudorandomIndexes(seed, count, max, excludeMultiplesOf = 0) {
-    const maxCount = excludeMultiplesOf ? max - max / excludeMultiplesOf : max;
-    if (maxCount < count)
-        throw Error(`Cannot select ${count} unique pseudorandom indexes from ${max} values`);
-    const maxIterations = BigInt(count * 1000);
-    const modulus = BigInt(max);
-    const skip = BigInt(excludeMultiplesOf);
-    const indexes = new Set();
-    const state = sha256(seed);
-    for (let i = 0n; i < maxIterations; i++) {
-        let index = sha256(state + i) % modulus;
-        if (skip && index % skip === 0n)
-            continue; // if the index should be excluded, skip it
-        if (indexes.has(index))
-            continue; // if the index is already in the list, skip it
-        indexes.add(index);
-        if (indexes.size >= count)
-            break; // if we have enough indexes, break the loop
-    }
-    // if we couldn't generate enough indexes within max iterations, throw an error
-    if (indexes.size < count)
-        throw new Error(`Could not generate ${count} pseudorandom indexes`);
-    const result = [];
-    for (let index of indexes) {
-        result.push(Number.parseInt(index.toString(16), 16));
-    }
-    return result;
-}
-exports.getPseudorandomIndexes = getPseudorandomIndexes;
 function vectorToBuffers(values, size) {
     const result = new Array(values.length);
     if (values.elementSize > size) {
@@ -80,14 +50,6 @@ function buffersToBigInts(values) {
     return result;
 }
 exports.buffersToBigInts = buffersToBigInts;
-function sha256(value) {
-    const buffer = (typeof value === 'bigint')
-        ? Buffer.from(value.toString(16), 'hex')
-        : value;
-    const hash = crypto.createHash('sha256').update(buffer);
-    return BigInt('0x' + hash.digest().toString('hex'));
-}
-exports.sha256 = sha256;
 function readBigInt(buffer, offset, elementSize) {
     const blocks = elementSize >> 3;
     let value = 0n;
