+++ examples/mimc/mimc256.ts
@@ -30,7 +30,7 @@ define MiMC over prime field (2^256 - 351 * 2^32 + 1) {
     using 1 readonly register {
         $k0: repeat [${roundConstants.join(', ')}];
     }
-}`);
+}`, { hashAlgorithm: 'wasmBlake2s256' });
 
 // TESTING
 // ================================================================================================
