+++ lib/components/index.ts
@@ -2,3 +2,4 @@ export { ZeroPolynomial } from './ZeroPolynomial';
 export { BoundaryConstraints } from './BoundaryConstraints';
 export { LinearCombination } from './LinearCombination';
 export { LowDegreeProver } from './LowDegreeProver';
+export { QueryIndexGenerator } from './QueryIndexGenerator';
